tools/requirements-qr.txt
qrcode[pil]
pillow
amzqr
